+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/ui/Window.java
@@ -51,7 +51,7 @@ public class Window extends Group implements Signal.Listener<KeyEvent> {
 	protected NinePatch chrome;
 
 	public static final int WHITE = 0xFFFFFF;
-	public static final int TITLE_COLOR = 0xFFFF44;
+	public static final int TITLE_COLOR = 0xE8C883;
 	public static final int SHPX_COLOR = 0x33BB33;
 	
 	public Window() {