+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/windows/WndTabbed.java
@@ -245,12 +245,15 @@ public class WndTabbed extends Window {
 		
 		protected Image icon;
 		private RectF defaultFrame;
+		private float logicalWidth, logicalHeight;
 		
 		public IconTab( Image icon ){
 			super();
 			
 			this.icon.copy(icon);
 			this.defaultFrame = icon.frame();
+			logicalWidth=icon.width;
+			logicalHeight=icon.height;
 		}
 		
 		@Override
@@ -266,6 +269,7 @@ public class WndTabbed extends Window {
 			super.layout();
 			
 			icon.frame(defaultFrame);
+			icon.logicalSize(logicalWidth,logicalHeight);
 			icon.x = x + (width - icon.width) / 2;
 			icon.y = y + (height - icon.height) / 2 - 1;
 			if (!selected) {
@@ -273,8 +277,10 @@ public class WndTabbed extends Window {
 				//if some of the icon is going into the window, cut it off
 				if (icon.y < y + CUT) {
 					RectF frame = icon.frame();
-					frame.top += (y + CUT - icon.y) / icon.texture.height;
+					float clipped=y+CUT-icon.y;
+					frame.top += clipped / logicalHeight * defaultFrame.height();
 					icon.frame( frame );
+					icon.logicalSize(logicalWidth,Math.max(0,logicalHeight-clipped));
 					icon.y = y + CUT;
 				}
 			}